Implement byte-format content matching
Sometimes users need to find the presence of content in a response
instead of just knowing if a URI exists. Sometimes, the content is
in a language using a character set other than UTF-8. Sometimes,
users need to do this very quickly and at scale. Now is one of
those times.

Implement a `Vec<Vec<u8>>` in the UriScanner struct to hold a set
of desired matches from content acquired by requests. Implement a
check for the presence of elements in that outer vector, which if
successful will attempt to find each element in the outer vector in
the bytes of the response (ignoring the response code produced).

Fix a logic flaw that allowed empty words to be added to the URI
list (`Vec<String>`).

Testing:
  Minimal through update to the uri_scanner example - doesn't blow
up yet, finds results, works as previously with no content matches
supplied.